requirements.txt
Django==2.0
django-bootstrap3==9.0.0
imdbpie==5.2.0
simplejson==3.11.1
imdbpy==6.6